Refactor to enable settings summary in sidebar
