+++ apps/ts-move-in-out-angular/src/styles.scss
@@ -1282,7 +1282,7 @@ button[type='submit'] {
   margin-bottom: 1rem;
 }
 
-.center {
+.center-text {
   text-align: center;
 }
