+++ libs/ts/movein/ngx/src/lib/modules/builder/modules/date-select/date-select.component.html
@@ -1,5 +1,5 @@
 <div class="container12">
-  <p class="center component-instructions">Select your move-in day.</p>
+  <p class="center-text component-instructions">Select your move-in day.</p>
 
   <div class="date-card-container">
     <div *ngFor="let item of timestampedDates; let i = index;" class="date-card" (click)="saveDate(item)" (keyup.space)="saveDate(item)" (keyup.enter)="saveDate(item)" tabindex="0">
