+++ libs/ts/movein/ngx/src/lib/modules/builder/modules/review/review.component.html
@@ -1,5 +1,5 @@
 <div class="container12">
-  <p class="trailer-2 center component-instructions">Please review that your selections are correct. In doing so, you'll receive the best parking locations and avoid parking citations on your move-in day.</p>
+  <p class="trailer-2 center-text component-instructions">Please review that your selections are correct. In doing so, you'll receive the best parking locations and avoid parking citations on your move-in day.</p>
 
   <div class="flex flex-row justify-center">
     <div class="column6 flex flex-column" id="settings-table">
