+++ libs/ts/movein/ngx/src/lib/modules/builder/modules/zone-select/zone-select.component.html
@@ -1,5 +1,5 @@
 <div class="container12">
-  <p class="center component-instructions">Select your residence hall.</p>
+  <p class="center-text component-instructions">Select your residence hall.</p>
 
   <ng-container *ngIf="savedResidence | async as residence">
     <div class="zone-container">
