+++ libs/ts/movein/ngx/src/lib/modules/map/services/move-in-out-settings/move-in-out-settings.service.ts
@@ -4,6 +4,7 @@ import { EnvironmentService } from '@tamu-gisc/common/ngx/environment';
 import { LocalStoreService } from '@tamu-gisc/common/ngx/local-store';
 
 import {
+  MoveDate,
   MoveDates,
   MoveEventType,
   MoveInSettings,
@@ -152,6 +153,79 @@ export class MoveInOutSettingsService {
     }
 
     return day;
+  }
+
+  /**
+   * Returns MoveDates for either move-in or move-out event.
+   */
+  public getDaysForType(type: MoveEventType) {
+    if (this.days && this.days[type] && this.days[type].length > 0) {
+      return this.days[type];
+    } else {
+      console.warn(`No move dates found for ${type} event.`);
+      return undefined;
+    }
+  }
+
+  /**
+   * Returns the move-in/out MoveDate object for the provided matching calendar day (find operation)
+   */
+  public getDateForDay(type: MoveEventType, day: number | string): MoveDate | undefined {
+    if (this.days[type]?.length > 0) {
+      return this.days[type].find((d) => d.day == day);
+    } else {
+      console.warn(`No move date for provided '${type}' type and '${day}' day.`);
+      return undefined;
+    }
+  }
+
+  /**
+   * Returns the first move in/out MoveDate object or the provided event type.
+   */
+  public getFirstMoveDate(type: MoveEventType) {
+    return this.getDaysForType(type)?.[0];
+  }
+
+  /**
+   * Accepts a move event type and a MoveDate object and determines the index of the `date` object relative to the
+   * event type's MoveDates array.
+   */
+  public getMoveDateEventDayNumber(type: MoveEventType, date: MoveDate) {
+    const dates = this.getDaysForType(type);
+
+    if (dates) {
+      return dates.findIndex((d) => d.day == date.day) + 1;
+    } else {
+      console.warn(`No move dates found for ${type} event.`);
+      return undefined;
+    }
+  }
+
+  public getMoveDateEventDayNumberForSettings() {
+    const savedDate = this.getDateForDay('in', parseInt(this.settings.date));
+
+    if (savedDate) {
+      return this.getMoveDateEventDayNumber('in', savedDate);
+    } else {
+      console.warn(`No move date found for settings date.`);
+      return undefined;
+    }
+  }
+
+  /**
+   * Returns the move-in/out date as a Date object for the provided MoveDate object.
+   *
+   * Optionally provide a date object. Settings date will be used if not provided.
+   */
+  public getMoveDateEventAsDate(type: MoveEventType, date?: MoveDate) {
+    const savedDate = this.getDateForDay(type, date ? date.day : parseInt(this.settings.date));
+
+    if (savedDate) {
+      return new Date(new Date().getFullYear(), savedDate.month - 1, savedDate.day);
+    } else {
+      console.warn(`No event day found for provided date.`);
+      return undefined;
+    }
   }
 
   private _validateResidence(buildingNumber: string) {
