+++ libs/ts/movein/ngx/src/lib/modules/sidebar/components/sidebar-reference/sidebar-reference.component.scss
@@ -1,0 +1,6 @@
+.settings-edit {
+  text-align: right;
+  font-size: 0.75rem;
+  margin-right: 1rem;
+  padding-bottom: 0.25rem;
+}
